🐛 fix(server.ts): change port variable case from lowercase port to uppercase PORT to improve semantics ✨ feat(server.ts): add support for process.env.PORT environment variable to be able to run app on a configurable port 🐛 fix(page.jsx): remove unnecessary console.log statement ✨ feat(page.jsx): add support for fetching 'virksomheder' data from the server 🐛 fix(globals.css): fix background and foreground colors for dark mode ✨ feat(globals.css): import 'Inter' font from Google Fonts and set it as the default font for the body 🐛 fix(layout.jsx): remove unused 'Inter' font import ✨ feat(layout.jsx): add metadata for title and description 🐛 fix(page.jsx): remove unused 'landing.css' import ✨ feat(page.jsx): add conditional rendering for the button based on session availability 🐛 fix(AddVirksomhed.jsx): remove unused 'tags' state and 'fetchTags' function ✨ feat(AddVirksomhed.jsx): add support for adding a new virksomhed to the database 🐛 fix(KontaktTable.jsx): remove unnecessary condition for accessing 'virksomhed' property ✨ feat(ManageVirksomhed.jsx): add component for managing virksomhed data 🐛 fix(VirksomhedInfo.jsx): remove unused 'tags' state ✨ feat(VirksomhedInfo.jsx): add support for editing virksomhed data 🐛 fix(VirksomhedRowManipulation.jsx): add condition to delete virksomheder only for the logged-in user ✨ feat(VirksomhedTable.jsx): add support for row manipulation and selection 🐛 fix(LandingPageHeader.jsx): change Sherpa to h2 element ✨ feat(next.config.js): enable experimental ppr feature 🐛 fix(package.json): update next version to 14.0.5-canary.42

+++ components/VirksomhedRowManipulation.jsx
@@ -12,6 +12,7 @@ function RowManipulation({
   setRowSelection,
   refreshData,
   selectedIds,
+  session,
 }) {
   const supabase = createClientComponentClient();
 
@@ -20,7 +21,9 @@ function RowManipulation({
     const { data, error } = await supabase
       .from("virksomheder")
       .delete()
-      .in("id", selectedIds);
+      .in("id", selectedIds)
+      .eq("user_id", session.user.id);
+
     if (error) {
       toast.error("Noget gik galt");
     } else {
